requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.0.3
gunicorn==20.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
setuptools==75.5.0
Werkzeug==3.0.6
